# Conflicts:
#	core-customize/hybris/bin/custom/gallagher/gallagherfacades/resources/gallagherfacades-spring.xml
#	core-customize/hybris/bin/custom/gallagherb2c/gallagherb2cstorefront/web/webroot/WEB-INF/tags/responsive/template/javaScript.tag

+++ core-customize/hybris/bin/custom/gallagher/gallaghercore/src/com/gallagher/core/services/impl/GallagherProductProcessingServiceImpl.java
@@ -298,18 +298,21 @@ public class GallagherProductProcessingServiceImpl implements GallagherProductPr
 		final List<ProductFeatureModel> newProductFeatures = new ArrayList<>();
 		final List<ProductFeatureModel> oldProductFeatures = variantProduct.getFeatures();
 
-		modelService.removeAll(oldProductFeatures);
-
-		for (final ProductFeatureModel productFeature : productFeatures)
+		if (!CollectionUtils.isEmpty(oldProductFeatures))
 		{
-			final ProductFeatureModel newProductFeature = modelService.clone(productFeature);
+			modelService.removeAll(oldProductFeatures);
 
-			newProductFeature.setProduct(variantProduct);
+			for (final ProductFeatureModel productFeature : productFeatures)
+			{
+				final ProductFeatureModel newProductFeature = modelService.clone(productFeature);
 
-			newProductFeatures.add(newProductFeature);
+				newProductFeature.setProduct(variantProduct);
+
+				newProductFeatures.add(newProductFeature);
+			}
+
+			modelService.saveAll(newProductFeatures);
 		}
-
-		modelService.saveAll(newProductFeatures);
 	}
 
 	/**
